requirements.txt
wheel
pycryptodome
six
python-decouple
requests~=2.24.0
django~=3.1.2
django-cors-headers
djangorestframework==3.11.2
pytz
sqlparse
coreapi
django-filter~=2.4.0
mysqlclient
redis
celery
django-redis
drf-yasg
sentry-sdk
python-dateutil
django-debug-toolbar~=3.1.1
packaging
django-json-widget

gunicorn
gevent
drf-spectacular

geoip2
drf-extensions